Adds remaining adapter.requestDevice tests.
- Removes duplicated tests, and consolidates the limits tests
